+++ src/Faker/Extensions/ArrayExtensions.cs
@@ -10,7 +10,6 @@ namespace RimuTec.Faker.Extensions {
       public static T Sample<T>(this T[] array) {
          if (array.Count() == 0) {
             return default(T);
-            //throw new InvalidOperationException($"Array {nameof(array)} must contain at least one item");
          }
 
          return array[RandomNumber.Next(0, array.Length)];